velociterm/helpers/auth_helper.py
# auth_helpers.py

from fastapi import HTTPException, Cookie, Depends, status
import jwt  # PyJWT
from datetime import datetime, timedelta
from uuid import uuid4
# JWT and Security Configuration
SECRET_KEY = "your-secret-key"
ALGORITHM = "HS256"

async def get_current_user(access_token: str = Cookie(None)):
    if not access_token:
        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Not authenticated")
    try:
        payload = jwt.decode(access_token, SECRET_KEY, algorithms=[ALGORITHM])
        username = payload.get("sub")
        if not username:
            raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Invalid token")
        return username
    except jwt.ExpiredSignatureError:
        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Token expired")
    except jwt.InvalidTokenError:
        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Invalid token")
