requirements.txt
annotated-types>=0.6.0
anyio>=4.2.0
bcrypt>=4.1.2
bottle>=0.12.25
certifi>=2024.8.30
cffi>=1.16.0
charset-normalizer>=3.4.0
click>=8.1.7
clr-loader>=0.2.6
colorama>=0.4.6
cryptography>=41.0.7
exceptiongroup>=1.2.0
fastapi>=0.109.0
fastapi-sessions>=0.3.2
h11>=0.14.0
httptools>=0.6.1
idna>=3.6
itsdangerous>=2.2.0
Jinja2>=3.1.3
ldap3>=2.9.1
MarkupSafe>=2.1.3
packaging>=24.1
paramiko>=3.2.0
passlib>=1.7.4
proxy_tools>=0.1.0
pyasn1>=0.6.1
pycparser>=2.21
pydantic>=2.5.3
pydantic_core>=2.14.6
pydantic_settings>=2.7.1
PyJWT>=2.9.0
PyNaCl>=1.5.0
pynetbox>=7.4.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
pythonnet>=3.0.3
pywebview>=4.4.1
PyYAML>=6.0.1
requests>=2.32.3
sniffio>=1.3.0
starlette>=0.35.1
typing_extensions>=4.9.0
urllib3>=2.2.3
uvicorn>=0.26.0
watchfiles>=0.21.0
websockets>=12.0
